fixed timetag issue in raw.cpp and other small bugs while attempting to analyze preliminary nickel data

+++ analysis/include/dataStructures.h
@@ -35,8 +35,7 @@ struct RawEvent
     unsigned int baseline;
     unsigned int flags;
     unsigned int extTime; // "extended timestamp of event" extends timetag with
-    // 16 additional bits for times greater than 2^32
-    // sample periods.
+    // 16 additional bits for times greater than 2^32 sample periods.
     double fineTime;// "fine timestamp of event" sub-divides timetag with
     // 10 additional bits of time granularity, with units of
     // (sample period)/2^10 units.
@@ -63,9 +62,9 @@ struct SeparatedEvent
 struct ProcessedEvent
 {
     unsigned int macroNo; // label each event by macropulse
-    double macroTime; // provide the macropulse "zero" time
+    long double macroTime; // provide the macropulse "zero" time
     unsigned int evtNo; // uniquely label each event in a macropulse
-    double completeTime; // the event's 48-bit timestamp
+    long double completeTime; // the event's 48-bit timestamp
     double fineTime; // the event's fine time
     unsigned int targetPos; // target position
     unsigned int sgQ, lgQ; // the event's short and long integrated charge gates
@@ -77,7 +76,7 @@ struct TargetChangerEvent
 {
     unsigned int macroNo; // label each event by macropulse
     unsigned int modeChange; // indicate the first event after a mode change
-    double macroTime; // the event's time-zero reference (the macropulse start)
+    long double macroTime; // the event's time-zero reference (the macropulse start)
     double fineTime; // the event's calculated fine time
     unsigned int targetPos; // target position
     unsigned int lgQ; // the event's long integrated charge gate
